# perf: Graphql CDN using Stellate