--- a/utils/constants.ts
+++ b/utils/constants.ts
@@ -55,7 +55,7 @@ export const URLS = {
     snek: 'https://squid.subsquid.io/snekk/graphql',
     snekRococo: 'https://squid.subsquid.io/sneck/graphql',
     marck: 'https://squid.subsquid.io/marck/graphql',
-    stick: 'https://squid.subsquid.io/stick/graphql',
+    stick: 'https://query-stick.stellate.sh',
     speck: 'https://squid.subsquid.io/speck/graphql',
     polkassembly: 'https://squid.subsquid.io/polkadot-polkassembly/graphql',
     gsPolkdot: 'https://squid.subsquid.io/gs-main-polkadot/graphql',