Hi @kodadot/internal-dev. Would you like to try Stellate for our GraphQL endpoint? In a simple comparison, it appears to be faster. This could help us if we get a large number of visitors at the same time. Perhaps if the number of visitors is not too high, the effect might not be too noticeable. For now, I’ve set the cache on this endpoint to last for 60 seconds.

| Before | After |
|---|---|
| ![Screenshot 2023-10-12 152306](https://user-images.githubusercontent.com/734428/274536686-b5627c30-70ce-439f-8839-0093b65cbfec.png) | ![Screenshot 2023-10-12 152243](https://user-images.githubusercontent.com/734428/274536700-55449b5b-8576-484a-b03a-804a9b915304.png) |
| https://canary.kodadot.xyz/ahk/collection/150 | https://deploy-preview-7623--koda-canary.netlify.app/ahk/collection/150 |


https://stellate.co/docs/graphql-edge-cache/cache-rules
```js
// cache rules

rules: [
  {
    types: [
      'NFTEntity',
      'NFTEntitiesConnection',
      'CollectionEntity',
      'CollectionEntitiesConnection',
      'Event',
      'EventEntity',
      'LastEventEntity'
    ],
    maxAge: 60,
    swr: 600,
    description: 'Cache some types',
  },
],
```

## PR Type

- [ ] Bugfix
- [x] Feature
- [ ] Refactoring

## Context

- [x] Related: https://github.com/kodadot/nft-gallery/issues/2743
- [x] Related: https://github.com/kodadot/nft-gallery/issues/4657

#### Had issue bounty label?

- [x] My DOT address: [Payout](https://canary.kodadot.xyz/dot/transfer/?target=1xjvRADwdJcnmUCLWerEHR4iGCT5EBTm4D4fzLLg4LcAC9p)

## Copilot Summary
<!--
copilot:summary
-->
### <samp>🤖 Generated by Copilot at 45af347</samp>

This pull request adds support for the new Stellate indexers, which are a different kind of Subsquid indexer for the NFT gallery. It updates the `indexers.ts` and `constants.ts` files to handle both Stellate and Squid indexers and use the appropriate one for each network.

<!--
copilot:poem
-->
### <samp>🤖 Generated by Copilot at 45af347</samp>

> _`SquidEndpoint` changes_
> _Stellate indexers join in_
> _Autumn of networks_