@preschian 2023-10-12
> How does it work? does it only cache data?

It wraps the subsquid endpoint with stellate

```
before:
user -> https://squid.subsquid.io/stick/graphql

after:
user -> https://query-stick.stellate.sh/ -> cache miss -> https://squid.subsquid.io/stick/graphql
user -> https://query-stick.stellate.sh/ -> cache hit -> return the cache response
```

> Will we need it for all indexers? wouldn't we have a problem on the billing?

for testing maybe we can test with one endpoint first? they have a free plan with unlimited requests. the limitation is only working on 2 endpoints, and need to put their badge

- https://stellate.co/pricing
- https://stellate.co/docs/platform/adding-powered-by-stellate-badge

> Do you have acces at some sort of backend dashboard?

I just created the account by myself. what we need only subsquid endpoint to use stellate