@Jarsen136 2023-10-12
Caching graphql is a good idea, but I also have some concerns.

- I wonder if it supports bypassing the cache. For example, if a user buy a nft we need to fetch the latest information instead of cache.

- Is it compatible with the current real-time subscription feature `useSubscriptionGraphql`? I'm afraid the cache would affect it.