@preschian 2023-10-12
> Caching graphql is a good idea, but I also have some concerns.
> 
> * I wonder if it supports bypassing the cache. For example, if a user buy a nft we need to fetch the latest information as fast as possible instead of cache.
> * Is it compatible with the current real-time subscription feature `useSubscriptionGraphql`? I'm afraid the cache would affect it.

we can use custom header to bypass the cache https://stellate.co/docs/graphql-edge-cache/bypass-the-cache

or, not putting all Types in the cache. currently in the rules I put 

```js
[
  'NFTEntity',
  'NFTEntitiesConnection',
  'CollectionEntity',
  'CollectionEntitiesConnection',
  'Event',
  'EventEntity',
  'LastEventEntity',
]
```