@yangwao 2023-10-12
Ah GraphCDN => Stellate?

We've been considering this one in the past.