@preschian 2023-10-12
For the subscription query, what if we fallback to the subsquid endpoint? like this c01503c20eed8db3b2dc55266c1bba17c4551f8b



> Ah GraphCDN => Stellate?
> 
> We've been considering this one in the past.

yes, cdn for graphql