@vikiival 2023-10-26
So looking forward to hear opinion from @kodadot/internal-dev about using stellate.

> unfortunately, we need to adjust the subscription over HTTP instead of WS

We can fallback to the sqd endpoint. 
the point why we use WebSocket on some parts of the app is the reason that we need fresh data.

But I would be happy to try it on AssetHubs (2 endpoints) and if we would see potential benefit we can scale from there.