@roiLeo 2023-10-26
> So looking forward to hear opinion from @kodadot/internal-dev about using stellate.

meh.. I have no opinion

> But I would be happy to try it on AssetHubs (2 endpoints) and if we would see potential benefit we can scale from there.

yes we could try that first