@preschian 2023-10-30
lets try it then? @vikiival 
added Stellate badge in the footer

![image](https://github.com/kodadot/nft-gallery/assets/734428/e50d7d3d-fcad-4e34-af88-121c65dba441)