@vikiival 2023-10-30
> added Stellate badge in the footer

Shouldn't be the badge in the readme?