@yangwao 2023-11-02
😍 Perfect, I’ve sent the payout 
💵 $50 @ 4.58 USD/DOT ~ 10.917 $DOT
🧗 [1xjvRADwdJcnmUCLWerEHR4iGCT5EBTm4D4fzLLg4LcAC9p](https://polkadot.subscan.io/account/1xjvRADwdJcnmUCLWerEHR4iGCT5EBTm4D4fzLLg4LcAC9p)
🔗 [0x3a9f96740a41ae5e790748c4f0b85d17e381cf7562d08bb4aa24e3880fd72967](https://polkadot.subscan.io/extrinsic/0x3a9f96740a41ae5e790748c4f0b85d17e381cf7562d08bb4aa24e3880fd72967)

🪅 Let’s grab another issue and get rewarded! 
🪄 [github.com/kodadot/nft-gallery/issues](https://github.com/kodadot/nft-gallery/issues/)