@yangwao 2023-11-02
looking for improvement. 
300ms cut seems fair, might improve the lighthouse?