@preschian 2023-11-03
> might improve the lighthouse?

It is probably not significant since I only set the cache to 60 seconds
but if the LCP in some page rely on graphql request, we can test to set a longer cache on that