@vikiival 2023-11-03
> but if the LCP in some page rely on graphql request,

90% of landing 🥹