@preschian 2023-11-03
> 90% of landing 🥹

aahh, I see. let me check