@preschian 2023-11-03
After some checking, our FCP is currently a bit high. Currently, adjustment on LCP will not give significant results. We need to reduce the javascript size on the client

![image](https://github.com/kodadot/nft-gallery/assets/734428/687daee0-7885-41e3-a816-678a15217179)

SSR or reducing unused JS in the pages could be helpful in this case, I guess 

![image](https://github.com/kodadot/nft-gallery/assets/734428/bf380a8c-57d6-447b-9afd-8a12202a5a0b)