@preschian 2023-11-03
> > SSR or reducing unused JS in the pages could be helpful in this case, I guess
> 
> feel free to test on #7725

checking 👍