@preschian 2023-11-03
confirmed, with SSR from #7725 looks better

FCP from 9xx ms -> 6xx ms
Unused JS 3.2 MB -> 2.5 MB

this will help us to increase the scores

![image](https://github.com/kodadot/nft-gallery/assets/734428/92df2fab-21d3-41dc-a539-5331b563ecf4)